fixed : decimals values not matching with UI

+++ src/main/java/com/actualize/mortgage/utils/Convertor.java
@@ -679,7 +679,7 @@ public class Convertor {
 	{
 		if(checkAmountNotNull(amount))
 		{
-			float val = Float.parseFloat(amount);
+			double val = Double.parseDouble(amount);
 			return String.format("%.2f", val);
 		}
 		return amount;
